Fixed bug in return type of compound assignment
The wrong return type of the compound arithmetic operators defined in
data_types.h was causing LLVM to throw illegal instruction SIGILL. Why
this is happening is not clear, as the bug doesn't seem to be so
distruptive. To replicate the bug, it is sufficient to revert the
change at line 230 and compile with -O0.